+++ src/main/java/kitchenpos/domain/menu_group/MenuGroup.java
@@ -1,19 +1,17 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu_group;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigDecimal;
 
 @Entity
-public class Product {
+public class MenuGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private BigDecimal price;
 
     public Long getId() {
         return id;
@@ -29,13 +27,5 @@ public class Product {
 
     public void setName(final String name) {
         this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
     }
 }
