+++ src/main/java/kitchenpos/domain/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu_group.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
