+++ src/main/java/kitchenpos/domain/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.repository;
 
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
