+++ src/main/java/kitchenpos/domain/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
