+++ src/test/java/support/fixture/TableBuilder.java
@@ -1,7 +1,7 @@
 package support.fixture;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order_table.OrderTable;
+import kitchenpos.domain.table_group.TableGroup;
 
 public class TableBuilder {
